cleaned up AL notebook and renamed files
